chore: remove dead BridgeTokenSelectorBase component
Extract SkeletonItem into its own file and delete BridgeTokenSelectorBase,
which was never used as a component (only SkeletonItem was imported from it).
