+++ app/components/UI/Bridge/components/BridgeTokenSelector/BridgeTokenSelector.test.tsx
@@ -293,7 +293,7 @@ jest.mock(
   },
 );
 
-jest.mock('../BridgeTokenSelectorBase', () => ({
+jest.mock('../SkeletonItem', () => ({
   SkeletonItem: () => {
     const { createElement } = jest.requireActual('react');
     const { View } = jest.requireActual('react-native');
